move subteamtask class to file in models folder
