+++ models/subteamtask.py
@@ -1,0 +1,8 @@
+class SubTeamTask:
+    def __init__(self, event_reference, creation_date, task_description, task_priority, assigned_by, assigned_to):
+        self.event_reference = event_reference
+        self.creation_date=creation_date
+        self.task_description=task_description
+        self.posittask_priorityion=task_priority
+        self.assigned_by=assigned_by
+        self.assigned_to=assigned_to
